lmb-1331: fetch burgemeesters and aangewezen burgemeesters when he mandatarissen are fetched
