+++ app/components/verkiezingen/burgemeester-alert.hbs
@@ -1,9 +1,7 @@
-<div {{did-insert (perform this.handleErrorMessage)}}>
-  {{#if this.errorMessage}}
-    <div class="au-o-box">
-      <AuAlert @skin="warning" @icon="alert-triangle" @closable={{true}}>
-        <p>{{this.errorMessage}}</p>
-      </AuAlert>
-    </div>
-  {{/if}}
-</div>
+{{#if this.errorMessage}}
+  <div class="au-o-box">
+    <AuAlert @skin="warning" @icon="alert-triangle" @closable={{true}}>
+      <p>{{this.errorMessage}}</p>
+    </AuAlert>
+  </div>
+{{/if}}
